+++ pycrunch_tests/dogfood/test_class_base.py
