+++ pycrunch_tests/dogfood/test_double_inheritance.py
@@ -1,4 +1,4 @@
-from pycrunch.tests.dogfood.helper_file import AbstractMyTestCase
+from pycrunch_tests.dogfood.helper_file import AbstractMyTestCase
 
 class DoublyInheritedScenario(AbstractMyTestCase):
     def test_1(self):
