+++ pycrunch_tests/test_auto_configuration.py
