+++ pycrunch_tests/test_insights.py
